Re-organize UI files
