+++ Plugin/Source/GUI/Assets/gui_ios.xml
@@ -30,12 +30,12 @@
   <View id="root" resizable="1" resize-corner="1" flex-direction="column"
         padding="0" width="580" height="580" background-color="FF8B3232"
         background-image="Background_svg" image-placement="stretch">
-    <View max-height="100" padding="0" margin="0" background-color="">
+    <View max-height="80" padding="0" margin="0" background-color="">
       <View margin="2" padding="" background-color="00000000" flex-direction="column"
             flex-grow="0.75">
         <View flex-grow="0.333" background-color="00000000"/>
         <TitleComp background-color="00000000" title="Chow Tape Model" font="28"
-                   padding="3" flex-grow="1.0"/>
+                   padding="1" flex-grow="1.0"/>
         <InfoComp background-color="00000000" text1="FFEAA92C" flex-grow="0.7"
                   padding="0" margin="5" border=""/>
         <View background-color="00000000" flex-grow="0.33"/>
