+++ Plugin/Source/GUI/OnOff/OnOffManager.cpp
