+++ Plugin/Source/GUI/OnOff/OnOffManager.h
