+++ Plugin/Source/GUI/Visualizers/LightMeter.cpp
