+++ Plugin/Source/GUI/Visualizers/LightMeter.h
