+++ Plugin/Source/GUI/Visualizers/MixGroupViz.h
