+++ Plugin/Source/GUI/Visualizers/TapeScope.cpp
