+++ Plugin/Source/GUI/Visualizers/TapeScope.h
